Modifies package.json to valid format so Autodesk registry would generate metadata.
The `devDependencies` in the package.json was an array which the Autodesk
npm registry does not like. It would fail to parse the metadata of the package
which meant it could not be detected for installation. Changing it to the
expected object with version numbers fixes the issue on Autodesk's registry.

Updated the version to 0.0.4